Fix minor invitation bugs & split vote fetch from movie suggestions

+++ Controllers/VoteController.cs
@@ -16,7 +16,57 @@ public class VoteController : ControllerBase
         _dbContext = dbContext;
     }
 
-    [HttpPost("suggestions/{movieSuggestionId}/vote")]
+    [HttpGet("movie-suggestions/vote/{voteId}")]
+    public async Task<IActionResult> GetVoteDetails([FromRoute] Guid voteId)
+    {
+        var vote = await _dbContext.Votes
+                        .Where(v => v.Id == voteId)
+                        .Select(v => new
+                        {
+                            Id = v.Id,
+                            VoteType = v.VoteType,
+                            MovieSuggestionId = v.MovieSuggestionId,
+                            MovieNightEventId = v.MovieSuggestion.MovieNightEventId,
+                            user = new
+                            {
+                                FullName = v.User.FullName,
+                                Email = v.User.Email
+                            }
+                        }) 
+                        .SingleOrDefaultAsync();
+        
+        return Ok(new {vote});
+    }
+
+    [HttpGet("movie-suggestions/{movieSuggestionId}/votes")]
+    public async Task<IActionResult> GetSuggestionVotes([FromRoute] string movieSuggestionId)
+    {
+ 
+        if (!Guid.TryParse(movieSuggestionId, out Guid parsedMovieSuggestionId))
+        {
+            return BadRequest(new CustomError { Message = "Invalid movieSuggestionId provided" });
+        }
+
+        var votes = await _dbContext.Votes
+            .Where(v => v.MovieSuggestionId == parsedMovieSuggestionId)
+            .Select(v => new
+            {
+                Id = v.Id,
+                VoteType = v.VoteType,
+                MovieSuggestionId = v.MovieSuggestionId,
+                MovieNightEventId = v.MovieSuggestion.MovieNightEventId,
+                user = new
+                {
+                    FullName = v.User.FullName,
+                    Email = v.User.Email
+                }
+            }) 
+            .ToListAsync();
+ 
+        return Ok( new { votes } );
+    }
+
+    [HttpPost("movie-suggestions/{movieSuggestionId}/vote")]
     public async Task<IActionResult> VoteForSuggestion([FromRoute] string movieSuggestionId, [FromBody] CreateVoteDto createDto)
     {
         if (!Guid.TryParse(movieSuggestionId, out Guid parsedMovieSuggestionId))
@@ -38,13 +88,38 @@ public class VoteController : ControllerBase
         var userExist = await _dbContext.Users.FindAsync(parsedUserId);
         if (userExist == null)
         {
-            return NotFound(new CustomError{ Message = "The user creating the movie night does not exist" });
+            return NotFound(new CustomError{ Message = "The user voting for a suggestion does not exist" });
         }
 
         // Delete other votes made by the user for the suggestion before adding a new one
-        await _dbContext.Votes 
+        var initialVote = await _dbContext.Votes
                 .Where(v => v.UserId == parsedUserId && v.MovieSuggestionId == parsedMovieSuggestionId)
-                .ExecuteDeleteAsync();
+                .AsNoTracking()
+                .FirstOrDefaultAsync();
+       
+        // If vote was found & its the same vote type, just remove it
+        if (initialVote != null && initialVote.VoteType == createDto.VoteType)
+        {
+            await _dbContext.Votes 
+                    .Where(v => v.UserId == parsedUserId && v.MovieSuggestionId == parsedMovieSuggestionId)
+                    .ExecuteDeleteAsync();
+            
+            if (createDto.VoteType == VoteType.Veto && movieSuggestionExists.IsDisqualified)
+            {
+                movieSuggestionExists.IsDisqualified = false;
+                await _dbContext.SaveChangesAsync();
+            }
+
+            return Ok(new { Message = "Vote removed successfully" });
+        }
+
+        // If voteType is different, delete the previous one before adding a new one
+        if (initialVote != null && initialVote.VoteType != createDto.VoteType)
+        {
+            await _dbContext.Votes 
+                    .Where(v => v.UserId == parsedUserId && v.MovieSuggestionId == parsedMovieSuggestionId)
+                    .ExecuteDeleteAsync();
+        }
 
         var vote = new Vote
         {
@@ -55,8 +130,17 @@ public class VoteController : ControllerBase
 
         await _dbContext.Votes.AddAsync(vote);
 
+        // If vote was a veto, disqualify the suggestion
+        if (createDto.VoteType == VoteType.Veto)
+        {
+            movieSuggestionExists.IsDisqualified = true;
+        } else if (createDto.VoteType != VoteType.Veto && movieSuggestionExists.IsDisqualified)
+        {
+            movieSuggestionExists.IsDisqualified = false;
+        }
+
         await _dbContext.SaveChangesAsync();
 
-        return Ok(new { Message = "Vote created successfully" });
+        return Ok(new { Message = "Vote added successfully" });
     }
 }
